Multilingual OCR/requirements.txt
surya-ocr
torch
transformers
pdf2image
Pillow
numpy
reportlab
gradio
